// src/modules/posts/posts.interface.ts
import { PostStatus } from "../../../generated/prisma/enums";
import { PostWhereInput } from "../../../generated/prisma/models";

export interface ICreatePostPayload {
  title: string;
  content: string;
  thumbnail?: string;
  isFeatured?: boolean;
  status?: PostStatus;
  tags: string[];
}


export interface IUpdatePostPayload {
     title: string;
  content: string;
  thumbnail?: string;
  isFeatured?: boolean;
  status?: PostStatus;
  tags: string[];
}

export interface IPostQuery extends PostWhereInput {
    // title?:string,
    // content?:string,
    
    searchTerm?:string,

    page?:string,
    limit?:string,
    sortOrder?:string,
    sortBy?:string
}

// enum Status {
//     DRAFT,
//   PUBLISHED,
//   ARCHIVED
// }